lint, minor refactoring
