+++ pkg/isbsvc/jetstream_service.go
@@ -26,6 +26,7 @@ import (
 	jsclient "github.com/numaproj/numaflow/pkg/shared/clients/nats"
 	"github.com/numaproj/numaflow/pkg/shared/logging"
 	"github.com/numaproj/numaflow/pkg/watermark/fetch"
+	"github.com/numaproj/numaflow/pkg/watermark/processor"
 	"github.com/numaproj/numaflow/pkg/watermark/store"
 	"github.com/numaproj/numaflow/pkg/watermark/store/jetstream"
 	"github.com/spf13/viper"
@@ -290,7 +291,7 @@ func (jss *jetStreamSvc) CreateWatermarkFetcher(ctx context.Context, bucketName 
 		return nil, err
 	}
 	storeWatcher := store.BuildWatermarkStoreWatcher(hbWatch, otWatch)
-	pm := fetch.NewProcessorManager(ctx, storeWatcher)
+	pm := processor.NewProcessorManager(ctx, storeWatcher)
 	watermarkFetcher := fetch.NewEdgeFetcher(ctx, bucketName, storeWatcher, pm)
 	return watermarkFetcher, nil
 }
