+++ pkg/watermark/fetch/edge_fetcher_test.go
@@ -18,10 +18,16 @@ package fetch
 
 import (
 	"context"
+	"fmt"
 	"strconv"
+	"sync"
 	"testing"
 	"time"
 
+	"github.com/nats-io/nats.go"
+	natstest "github.com/numaproj/numaflow/pkg/shared/clients/nats/test"
+	"github.com/numaproj/numaflow/pkg/watermark/store/inmem"
+	"github.com/numaproj/numaflow/pkg/watermark/store/jetstream"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap/zaptest"
 
@@ -41,11 +47,11 @@ func TestBuffer_GetWatermark(t *testing.T) {
 	hbWatcher := noop.NewKVOpWatch()
 	otWatcher := noop.NewKVOpWatch()
 	storeWatcher := store.BuildWatermarkStoreWatcher(hbWatcher, otWatcher)
-	processorManager := NewProcessorManager(ctx, storeWatcher)
+	processorManager := processor.NewProcessorManager(ctx, storeWatcher)
 	var (
-		testPod0     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 		pod0Timeline = []wmb.WMB{
 			{Watermark: 11, Offset: 9},
 			{Watermark: 12, Offset: 20},
@@ -68,24 +74,24 @@ func TestBuffer_GetWatermark(t *testing.T) {
 	)
 
 	for _, watermark := range pod0Timeline {
-		testPod0.offsetTimeline.Put(watermark)
+		testPod0.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod1Timeline {
-		testPod1.offsetTimeline.Put(watermark)
+		testPod1.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod2Timeline {
-		testPod2.offsetTimeline.Put(watermark)
+		testPod2.OffsetTimeline().Put(watermark)
 	}
-	processorManager.addProcessor("testPod0", testPod0)
-	processorManager.addProcessor("testPod1", testPod1)
-	processorManager.addProcessor("testPod2", testPod2)
+	processorManager.AddProcessor("testPod0", testPod0)
+	processorManager.AddProcessor("testPod1", testPod1)
+	processorManager.AddProcessor("testPod2", testPod2)
 
 	type args struct {
 		offset int64
 	}
 	tests := []struct {
 		name             string
-		processorManager *ProcessorManager
+		processorManager *processor.ProcessorManager
 		args             args
 		want             int64
 	}{
@@ -158,8 +164,8 @@ func Test_edgeFetcher_GetHeadWatermark(t *testing.T) {
 		hbWatcher         = noop.NewKVOpWatch()
 		otWatcher         = noop.NewKVOpWatch()
 		storeWatcher      = store.BuildWatermarkStoreWatcher(hbWatcher, otWatcher)
-		processorManager1 = NewProcessorManager(ctx, storeWatcher)
-		processorManager2 = NewProcessorManager(ctx, storeWatcher)
+		processorManager1 = processor.NewProcessorManager(ctx, storeWatcher)
+		processorManager2 = processor.NewProcessorManager(ctx, storeWatcher)
 	)
 
 	getHeadWMTest1(ctx, processorManager1)
@@ -167,7 +173,7 @@ func Test_edgeFetcher_GetHeadWatermark(t *testing.T) {
 
 	tests := []struct {
 		name             string
-		processorManager *ProcessorManager
+		processorManager *processor.ProcessorManager
 		want             int64
 	}{
 		{
@@ -195,11 +201,11 @@ func Test_edgeFetcher_GetHeadWatermark(t *testing.T) {
 	}
 }
 
-func getHeadWMTest1(ctx context.Context, processorManager1 *ProcessorManager) {
+func getHeadWMTest1(ctx context.Context, processorManager1 *processor.ProcessorManager) {
 	var (
-		testPod0     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 		pod0Timeline = []wmb.WMB{
 			{
 				Idle:      true,
@@ -224,24 +230,24 @@ func getHeadWMTest1(ctx context.Context, processorManager1 *ProcessorManager) {
 	)
 
 	for _, watermark := range pod0Timeline {
-		testPod0.offsetTimeline.Put(watermark)
+		testPod0.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod1Timeline {
-		testPod1.offsetTimeline.Put(watermark)
+		testPod1.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod2Timeline {
-		testPod2.offsetTimeline.Put(watermark)
+		testPod2.OffsetTimeline().Put(watermark)
 	}
-	processorManager1.addProcessor("testPod0", testPod0)
-	processorManager1.addProcessor("testPod1", testPod1)
-	processorManager1.addProcessor("testPod2", testPod2)
+	processorManager1.AddProcessor("testPod0", testPod0)
+	processorManager1.AddProcessor("testPod1", testPod1)
+	processorManager1.AddProcessor("testPod2", testPod2)
 }
 
-func getHeadWMTest2(ctx context.Context, processorManager1 *ProcessorManager) {
+func getHeadWMTest2(ctx context.Context, processorManager1 *processor.ProcessorManager) {
 	var (
-		testPod0     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 		pod0Timeline = []wmb.WMB{
 			{
 				Idle:      false,
@@ -266,17 +272,17 @@ func getHeadWMTest2(ctx context.Context, processorManager1 *ProcessorManager) {
 	)
 
 	for _, watermark := range pod0Timeline {
-		testPod0.offsetTimeline.Put(watermark)
+		testPod0.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod1Timeline {
-		testPod1.offsetTimeline.Put(watermark)
+		testPod1.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod2Timeline {
-		testPod2.offsetTimeline.Put(watermark)
+		testPod2.OffsetTimeline().Put(watermark)
 	}
-	processorManager1.addProcessor("testPod0", testPod0)
-	processorManager1.addProcessor("testPod1", testPod1)
-	processorManager1.addProcessor("testPod2", testPod2)
+	processorManager1.AddProcessor("testPod0", testPod0)
+	processorManager1.AddProcessor("testPod1", testPod1)
+	processorManager1.AddProcessor("testPod2", testPod2)
 }
 
 func Test_edgeFetcher_GetHeadWMB(t *testing.T) {
@@ -286,10 +292,10 @@ func Test_edgeFetcher_GetHeadWMB(t *testing.T) {
 		hbWatcher         = noop.NewKVOpWatch()
 		otWatcher         = noop.NewKVOpWatch()
 		storeWatcher      = store.BuildWatermarkStoreWatcher(hbWatcher, otWatcher)
-		processorManager1 = NewProcessorManager(ctx, storeWatcher)
-		processorManager2 = NewProcessorManager(ctx, storeWatcher)
-		processorManager3 = NewProcessorManager(ctx, storeWatcher)
-		processorManager4 = NewProcessorManager(ctx, storeWatcher)
+		processorManager1 = processor.NewProcessorManager(ctx, storeWatcher)
+		processorManager2 = processor.NewProcessorManager(ctx, storeWatcher)
+		processorManager3 = processor.NewProcessorManager(ctx, storeWatcher)
+		processorManager4 = processor.NewProcessorManager(ctx, storeWatcher)
 	)
 
 	getHeadWMBTest1(ctx, processorManager1)
@@ -299,7 +305,7 @@ func Test_edgeFetcher_GetHeadWMB(t *testing.T) {
 
 	tests := []struct {
 		name             string
-		processorManager *ProcessorManager
+		processorManager *processor.ProcessorManager
 		want             wmb.WMB
 	}{
 		{
@@ -341,11 +347,11 @@ func Test_edgeFetcher_GetHeadWMB(t *testing.T) {
 	}
 }
 
-func getHeadWMBTest1(ctx context.Context, processorManager1 *ProcessorManager) {
+func getHeadWMBTest1(ctx context.Context, processorManager1 *processor.ProcessorManager) {
 	var (
-		testPod0     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 		pod0Timeline = []wmb.WMB{
 			{
 				Idle:      true,
@@ -370,24 +376,24 @@ func getHeadWMBTest1(ctx context.Context, processorManager1 *ProcessorManager) {
 	)
 
 	for _, watermark := range pod0Timeline {
-		testPod0.offsetTimeline.Put(watermark)
+		testPod0.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod1Timeline {
-		testPod1.offsetTimeline.Put(watermark)
+		testPod1.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod2Timeline {
-		testPod2.offsetTimeline.Put(watermark)
+		testPod2.OffsetTimeline().Put(watermark)
 	}
-	processorManager1.addProcessor("testPod0", testPod0)
-	processorManager1.addProcessor("testPod1", testPod1)
-	processorManager1.addProcessor("testPod2", testPod2)
+	processorManager1.AddProcessor("testPod0", testPod0)
+	processorManager1.AddProcessor("testPod1", testPod1)
+	processorManager1.AddProcessor("testPod2", testPod2)
 }
 
-func getHeadWMBTest2(ctx context.Context, processorManager1 *ProcessorManager) {
+func getHeadWMBTest2(ctx context.Context, processorManager1 *processor.ProcessorManager) {
 	var (
-		testPod0     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 		pod0Timeline = []wmb.WMB{
 			{
 				Idle:      false,
@@ -412,24 +418,24 @@ func getHeadWMBTest2(ctx context.Context, processorManager1 *ProcessorManager) {
 	)
 
 	for _, watermark := range pod0Timeline {
-		testPod0.offsetTimeline.Put(watermark)
+		testPod0.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod1Timeline {
-		testPod1.offsetTimeline.Put(watermark)
+		testPod1.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod2Timeline {
-		testPod2.offsetTimeline.Put(watermark)
+		testPod2.OffsetTimeline().Put(watermark)
 	}
-	processorManager1.addProcessor("testPod0", testPod0)
-	processorManager1.addProcessor("testPod1", testPod1)
-	processorManager1.addProcessor("testPod2", testPod2)
+	processorManager1.AddProcessor("testPod0", testPod0)
+	processorManager1.AddProcessor("testPod1", testPod1)
+	processorManager1.AddProcessor("testPod2", testPod2)
 }
 
-func getHeadWMBTest3(ctx context.Context, processorManager1 *ProcessorManager) {
+func getHeadWMBTest3(ctx context.Context, processorManager1 *processor.ProcessorManager) {
 	var (
-		testPod0     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2     = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2     = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 		pod0Timeline = []wmb.WMB{
 			{
 				Idle:      false,
@@ -454,26 +460,556 @@ func getHeadWMBTest3(ctx context.Context, processorManager1 *ProcessorManager) {
 	)
 
 	for _, watermark := range pod0Timeline {
-		testPod0.offsetTimeline.Put(watermark)
+		testPod0.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod1Timeline {
-		testPod1.offsetTimeline.Put(watermark)
+		testPod1.OffsetTimeline().Put(watermark)
 	}
 	for _, watermark := range pod2Timeline {
-		testPod2.offsetTimeline.Put(watermark)
+		testPod2.OffsetTimeline().Put(watermark)
 	}
-	processorManager1.addProcessor("testPod0", testPod0)
-	processorManager1.addProcessor("testPod1", testPod1)
-	processorManager1.addProcessor("testPod2", testPod2)
+	processorManager1.AddProcessor("testPod0", testPod0)
+	processorManager1.AddProcessor("testPod1", testPod1)
+	processorManager1.AddProcessor("testPod2", testPod2)
 }
 
-func getHeadWMBTest4(ctx context.Context, processorManager1 *ProcessorManager) {
+func getHeadWMBTest4(ctx context.Context, processorManager1 *processor.ProcessorManager) {
 	var (
-		testPod0 = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-		testPod1 = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
-		testPod2 = NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
+		testPod0 = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
+		testPod1 = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod2"), 5)
+		testPod2 = processor.NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod3"), 5)
 	)
-	processorManager1.addProcessor("testPod0", testPod0)
-	processorManager1.addProcessor("testPod1", testPod1)
-	processorManager1.addProcessor("testPod2", testPod2)
+	processorManager1.AddProcessor("testPod0", testPod0)
+	processorManager1.AddProcessor("testPod1", testPod1)
+	processorManager1.AddProcessor("testPod2", testPod2)
+}
+
+func otValueToBytes(offset int64, watermark int64, idle bool) ([]byte, error) {
+	otValue := wmb.WMB{
+		Offset:    offset,
+		Watermark: watermark,
+		Idle:      idle,
+	}
+	otValueByte, err := otValue.EncodeToBytes()
+	return otValueByte, err
+}
+
+func TestFetcherWithSameOTBucket_InMem(t *testing.T) {
+	var (
+		err          error
+		pipelineName       = "testFetch"
+		keyspace           = "fetcherTest"
+		hbBucketName       = keyspace + "_PROCESSORS"
+		otBucketName       = keyspace + "_OT"
+		epoch        int64 = 1651161600000
+		testOffset   int64 = 100
+		wg           sync.WaitGroup
+	)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+	hbStore, hbWatcherCh, err := inmem.NewKVInMemKVStore(ctx, pipelineName, hbBucketName)
+	assert.NoError(t, err)
+	defer hbStore.Close()
+	otStore, otWatcherCh, err := inmem.NewKVInMemKVStore(ctx, pipelineName, otBucketName)
+	assert.NoError(t, err)
+	defer otStore.Close()
+
+	otValueByte, err := otValueToBytes(testOffset, epoch, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	epoch += 60000
+
+	otValueByte, err = otValueToBytes(testOffset+5, epoch, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p2", otValueByte)
+	assert.NoError(t, err)
+
+	hbWatcher, err := inmem.NewInMemWatch(ctx, "testFetch", keyspace+"_PROCESSORS", hbWatcherCh)
+	assert.NoError(t, err)
+	otWatcher, err := inmem.NewInMemWatch(ctx, "testFetch", keyspace+"_OT", otWatcherCh)
+	assert.NoError(t, err)
+	storeWatcher := store.BuildWatermarkStoreWatcher(hbWatcher, otWatcher)
+	var processorManager = processor.NewProcessorManager(ctx, storeWatcher)
+	var fetcher = NewEdgeFetcher(ctx, "testBuffer", storeWatcher, processorManager)
+	// start p1 heartbeat for 3 loops
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 3; i++ {
+			err = hbStore.PutKV(ctx, "p1", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+		err = hbStore.DeleteKey(ctx, "p1")
+		assert.NoError(t, err)
+	}()
+
+	// start p2 heartbeat for 20 loops (20 seconds)
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 20; i++ {
+			err = hbStore.PutKV(ctx, "p2", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+	}()
+
+	allProcessors := processorManager.GetAllProcessors()
+	for len(allProcessors) != 2 {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected 2 processors, got %d: %s", len(allProcessors), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	assert.True(t, allProcessors["p1"].IsActive())
+	assert.True(t, allProcessors["p2"].IsActive())
+
+	// "p1" status becomes deleted after 3 loops
+	for !allProcessors["p1"].IsDeleted() {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 to be deleted: %s", ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	allProcessors = processorManager.GetAllProcessors()
+
+	_ = fetcher.GetWatermark(isb.SimpleStringOffset(func() string { return strconv.FormatInt(testOffset, 10) }))
+	allProcessors = processorManager.GetAllProcessors()
+	assert.Equal(t, 2, len(allProcessors))
+	assert.True(t, allProcessors["p1"].IsDeleted())
+	assert.True(t, allProcessors["p2"].IsActive())
+	// "p1" should be deleted after this GetWatermark offset=101
+	// because "p1" offsetTimeline's head offset=100, which is < inputOffset 103
+	_ = fetcher.GetWatermark(isb.SimpleStringOffset(func() string { return strconv.FormatInt(testOffset+3, 10) }))
+	allProcessors = processorManager.GetAllProcessors()
+	assert.Equal(t, 1, len(allProcessors))
+	assert.True(t, allProcessors["p2"].IsActive())
+
+	time.Sleep(time.Second)
+	// resume p1 after one second
+	// run for 5 loops
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 5; i++ {
+			err = hbStore.PutKV(ctx, "p1", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+	}()
+
+	// wait until p1 becomes active
+	time.Sleep(1 * time.Second)
+	allProcessors = processorManager.GetAllProcessors()
+	for !allProcessors["p1"].IsActive() {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 to be active: %s", ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+	assert.True(t, allProcessors["p1"].IsActive())
+	assert.True(t, allProcessors["p2"].IsActive())
+	// "p1" has been deleted from vertex.Processors
+	// so "p1" will be considered as a new processors with a new default offset timeline
+	_ = fetcher.GetWatermark(isb.SimpleStringOffset(func() string { return strconv.FormatInt(testOffset+1, 10) }))
+	p1 := processorManager.GetProcessor("p1")
+	assert.NotNil(t, p1)
+	assert.True(t, p1.IsActive())
+	assert.NotNil(t, p1.OffsetTimeline())
+	assert.Equal(t, int64(-1), p1.OffsetTimeline().GetHeadOffset())
+
+	// publish a new watermark 101
+	otValueByte, err = otValueToBytes(testOffset+1, epoch, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	// "p1" becomes inactive after 5 loops
+	for !allProcessors["p1"].IsInactive() {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 to be inactive: %s", ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	time.Sleep(time.Second)
+
+	// resume after one second
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 3; i++ {
+			err = hbStore.PutKV(ctx, "p1", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+	}()
+
+	allProcessors = processorManager.GetAllProcessors()
+	for len(allProcessors) != 2 {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected 2 processors, got %d: %s", len(allProcessors), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	// added 101 in the previous steps for p1, so the head should be 101 after resume
+	assert.Equal(t, int64(101), p1.OffsetTimeline().GetHeadOffset())
+	wg.Wait()
+	cancel()
+}
+
+func TestFetcherWithSameOTBucket(t *testing.T) {
+	var (
+		keyspace         = "fetcherTest"
+		epoch      int64 = 1651161600000
+		testOffset int64 = 100
+		wg         sync.WaitGroup
+	)
+
+	s := natstest.RunJetStreamServer(t)
+	defer natstest.ShutdownJetStreamServer(t, s)
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+
+	// connect to NATS
+	nc, err := natstest.JetStreamClient(t, s).Connect(context.TODO())
+	assert.NoError(t, err)
+	defer nc.Close()
+
+	// create JetStream Context
+	js, err := nc.JetStream(nats.PublishAsyncMaxPending(256))
+	assert.NoError(t, err)
+
+	// create heartbeat bucket
+	_, err = js.CreateKeyValue(&nats.KeyValueConfig{
+		Bucket:       keyspace + "_PROCESSORS",
+		Description:  fmt.Sprintf("[%s] heartbeat bucket", keyspace),
+		MaxValueSize: 0,
+		History:      0,
+		TTL:          0,
+		MaxBytes:     0,
+		Storage:      nats.MemoryStorage,
+		Replicas:     0,
+		Placement:    nil,
+	})
+	defer func() { _ = js.DeleteKeyValue(keyspace + "_PROCESSORS") }()
+	assert.NoError(t, err)
+
+	// create offset timeline bucket
+	_, err = js.CreateKeyValue(&nats.KeyValueConfig{
+		Bucket:       keyspace + "_OT",
+		Description:  "",
+		MaxValueSize: 0,
+		History:      2,
+		TTL:          0,
+		MaxBytes:     0,
+		Storage:      nats.MemoryStorage,
+		Replicas:     0,
+		Placement:    nil,
+	})
+	defer func() { _ = js.DeleteKeyValue(keyspace + "_OT") }()
+	assert.NoError(t, err)
+
+	defaultJetStreamClient := natstest.JetStreamClient(t, s)
+
+	// create hbStore
+	hbStore, err := jetstream.NewKVJetStreamKVStore(ctx, "testFetch", keyspace+"_PROCESSORS", defaultJetStreamClient)
+	assert.NoError(t, err)
+	defer hbStore.Close()
+
+	// create otStore
+	otStore, err := jetstream.NewKVJetStreamKVStore(ctx, "testFetch", keyspace+"_OT", defaultJetStreamClient)
+	assert.NoError(t, err)
+	defer otStore.Close()
+
+	// put values into otStore
+
+	// this first entry should not be in the offset timeline because we set the wmb bucket history to 2
+	otValueByte, err := otValueToBytes(testOffset, epoch+100, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	otValueByte, err = otValueToBytes(testOffset+1, epoch+200, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	otValueByte, err = otValueToBytes(testOffset+2, epoch+300, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	epoch += 60000
+
+	otValueByte, err = otValueToBytes(testOffset+5, epoch+500, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p2", otValueByte)
+	assert.NoError(t, err)
+
+	// create watchers for heartbeat and offset timeline
+	hbWatcher, err := jetstream.NewKVJetStreamKVWatch(ctx, "testFetch", keyspace+"_PROCESSORS", defaultJetStreamClient)
+	assert.NoError(t, err)
+	otWatcher, err := jetstream.NewKVJetStreamKVWatch(ctx, "testFetch", keyspace+"_OT", defaultJetStreamClient)
+	assert.NoError(t, err)
+	storeWatcher := store.BuildWatermarkStoreWatcher(hbWatcher, otWatcher)
+	processorManager := processor.NewProcessorManager(ctx, storeWatcher)
+	fetcher := NewEdgeFetcher(ctx, "testBuffer", storeWatcher, processorManager)
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 3; i++ {
+			err = hbStore.PutKV(ctx, "p1", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+		err = hbStore.DeleteKey(ctx, "p1")
+		assert.NoError(t, err)
+	}()
+
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		// run p2 for 20 seconds
+		for i := 0; i < 20; i++ {
+			err = hbStore.PutKV(ctx, "p2", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+	}()
+
+	allProcessors := processorManager.GetAllProcessors()
+	for len(allProcessors) != 2 {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected 2 processors, got %d: %s", len(allProcessors), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	for allProcessors["p1"].OffsetTimeline().Dump() != "[1651161600300:102] -> [1651161600200:101] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1]" {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 has the offset timeline [1651161600300:102] -> [1651161600200:101] -> [-1:-1]..., got %s: %s", allProcessors["p1"].OffsetTimeline().Dump(), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	assert.True(t, allProcessors["p1"].IsActive())
+	assert.True(t, allProcessors["p2"].IsActive())
+
+	// "p1" is deleted after 5 loops
+	for !allProcessors["p1"].IsDeleted() {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 to be deleted: %s", ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	allProcessors = processorManager.GetAllProcessors()
+	assert.Equal(t, 2, len(allProcessors))
+	assert.True(t, allProcessors["p1"].IsDeleted())
+	assert.True(t, allProcessors["p2"].IsActive())
+
+	_ = fetcher.GetWatermark(isb.SimpleStringOffset(func() string { return strconv.FormatInt(testOffset, 10) }))
+	allProcessors = processorManager.GetAllProcessors()
+	assert.Equal(t, 2, len(allProcessors))
+	assert.True(t, allProcessors["p1"].IsDeleted())
+	assert.True(t, allProcessors["p2"].IsActive())
+	// "p1" should be deleted after this GetWatermark offset=101
+	// because "p1" offsetTimeline's head offset=102, which is < inputOffset 103
+	_ = fetcher.GetWatermark(isb.SimpleStringOffset(func() string { return strconv.FormatInt(testOffset+3, 10) }))
+	allProcessors = processorManager.GetAllProcessors()
+	assert.Equal(t, 1, len(allProcessors))
+	assert.True(t, allProcessors["p2"].IsActive())
+
+	time.Sleep(time.Second)
+	// resume after one second
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 5; i++ {
+			err = hbStore.PutKV(ctx, "p1", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+	}()
+
+	// wait until p1 becomes active
+	time.Sleep(1 * time.Second)
+	allProcessors = processorManager.GetAllProcessors()
+	for !allProcessors["p1"].IsActive() {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 to be active: %s", ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	assert.True(t, allProcessors["p1"].IsActive())
+	assert.True(t, allProcessors["p2"].IsActive())
+	// "p1" has been deleted from vertex.Processors
+	// so "p1" will be considered as a new processors with a new default offset timeline
+	_ = fetcher.GetWatermark(isb.SimpleStringOffset(func() string { return strconv.FormatInt(testOffset+1, 10) }))
+	p1 := processorManager.GetProcessor("p1")
+	assert.NotNil(t, p1)
+	assert.True(t, p1.IsActive())
+	assert.NotNil(t, p1.OffsetTimeline())
+	assert.Equal(t, int64(-1), p1.OffsetTimeline().GetHeadOffset())
+
+	// publish a new watermark 103
+	otValueByte, err = otValueToBytes(testOffset+3, epoch+500, false)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	// "p1" becomes inactive after 5 loops
+	for !allProcessors["p1"].IsInactive() {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 to be inactive: %s", ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	time.Sleep(time.Second)
+
+	// resume after one second
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		var err error
+		for i := 0; i < 10; i++ {
+			err = hbStore.PutKV(ctx, "p1", []byte(fmt.Sprintf("%d", time.Now().Unix())))
+			assert.NoError(t, err)
+			time.Sleep(1 * time.Second)
+		}
+	}()
+
+	allProcessors = processorManager.GetAllProcessors()
+	for len(allProcessors) != 2 {
+		select {
+		case <-ctx.Done():
+			t.Fatalf("expected 2 processors, got %d: %s", len(allProcessors), ctx.Err())
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	// added 103 in the previous steps for p1, so the head should be 103 after resume
+	assert.Equal(t, int64(103), p1.OffsetTimeline().GetHeadOffset())
+
+	for allProcessors["p1"].OffsetTimeline().Dump() != "[1651161660500:103] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1]" {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 has the offset timeline [1651161660500:103] -> [-1:-1] -> [-1:-1] -> [-1:-1]..., got %s: %s", allProcessors["p1"].OffsetTimeline().Dump(), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	// publish an idle watermark: simulate reduce
+	otValueByte, err = otValueToBytes(106, epoch+600, true)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	// p1 should get the head offset watermark from p2
+	for allProcessors["p1"].OffsetTimeline().Dump() != "[IDLE 1651161660600:106] -> [1651161660500:103] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1]" {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("expected p1 has the offset timeline [IDLE 1651161660600:106] -> [1651161660500:103] -> [-1:-1] -> [-1:-1]..., got %s: %s", allProcessors["p1"].OffsetTimeline().Dump(), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	// publish an idle watermark: simulate map
+	otValueByte, err = otValueToBytes(107, epoch+700, true)
+	assert.NoError(t, err)
+	err = otStore.PutKV(ctx, "p1", otValueByte)
+	assert.NoError(t, err)
+
+	// p1 should get the head offset watermark from p2
+	for allProcessors["p1"].OffsetTimeline().Dump() != "[IDLE 1651161660700:107] -> [IDLE 1651161660600:106] -> [1651161660500:103] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1] -> [-1:-1]" {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				t.Fatalf("[IDLE 1651161660700:107] -> [IDLE 1651161660600:106] -> [1651161660500:103] -> ..., got %s: %s", allProcessors["p1"].OffsetTimeline().Dump(), ctx.Err())
+			}
+		default:
+			time.Sleep(1 * time.Millisecond)
+			allProcessors = processorManager.GetAllProcessors()
+		}
+	}
+
+	wg.Wait()
+	cancel()
 }
