+++ pkg/watermark/processor/options.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fetch
+package processor
 
 type processorManagerOptions struct {
 	// podHeartbeatRate uses second as time unit
