+++ pkg/watermark/processor/options_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fetch
+package processor
 
 import (
 	"testing"
@@ -22,14 +22,18 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestProcessorHeartbeat(t *testing.T) {
-	hb := NewProcessorHeartbeat()
-	hb.Put("pod1", 1)
-	assert.Equal(t, int64(1), hb.Get("pod1"))
-	hb.Put("pod1", 5)
-	assert.Equal(t, int64(5), hb.Get("pod1"))
-	hb.Put("pod2", 6)
-	assert.Equal(t, map[string]int64{"pod1": int64(5), "pod2": int64(6)}, hb.GetAll())
-	hb.Delete("pod1")
-	assert.Equal(t, map[string]int64{"pod2": int64(6)}, hb.GetAll())
+func TestOptions(t *testing.T) {
+	testOpts := []ProcessorManagerOption{
+		WithPodHeartbeatRate(10),
+		WithRefreshingProcessorsRate(15),
+	}
+	opts := &processorManagerOptions{
+		podHeartbeatRate:         5,
+		refreshingProcessorsRate: 5,
+	}
+	for _, opt := range testOpts {
+		opt(opts)
+	}
+	assert.Equal(t, int64(10), opts.podHeartbeatRate)
+	assert.Equal(t, int64(15), opts.refreshingProcessorsRate)
 }
