+++ pkg/watermark/processor/processor_heartbeat_test.go
@@ -14,20 +14,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fetch
+package processor
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
-	"github.com/numaproj/numaflow/pkg/watermark/processor"
 )
 
-func TestFromProcessor_setStatus(t *testing.T) {
-	var ctx = context.Background()
-	p := NewProcessorToFetch(ctx, processor.NewProcessorEntity("testPod1"), 5)
-	p.setStatus(_inactive)
-	assert.Equal(t, _inactive, p.status)
+func TestProcessorHeartbeat(t *testing.T) {
+	hb := NewProcessorHeartbeat()
+	hb.Put("pod1", 1)
+	assert.Equal(t, int64(1), hb.Get("pod1"))
+	hb.Put("pod1", 5)
+	assert.Equal(t, int64(5), hb.Get("pod1"))
+	hb.Put("pod2", 6)
+	assert.Equal(t, map[string]int64{"pod1": int64(5), "pod2": int64(6)}, hb.GetAll())
+	hb.Delete("pod1")
+	assert.Equal(t, map[string]int64{"pod2": int64(6)}, hb.GetAll())
 }
