+++ pkg/watermark/processor/processor_to_fetch.go
@@ -14,17 +14,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fetch
+package processor
 
 import (
 	"context"
 	"fmt"
 	"sync"
 
+	"github.com/numaproj/numaflow/pkg/watermark/timeline"
 	"go.uber.org/zap"
 
 	"github.com/numaproj/numaflow/pkg/shared/logging"
-	"github.com/numaproj/numaflow/pkg/watermark/processor"
 )
 
 type status int
@@ -50,11 +50,19 @@ func (s status) String() string {
 // ProcessorToFetch is the smallest unit of entity (from which we fetch data) that does inorder processing or contains inorder data.
 type ProcessorToFetch struct {
 	ctx            context.Context
-	entity         processor.ProcessorEntitier
+	entity         ProcessorEntitier
 	status         status
-	offsetTimeline *OffsetTimeline
+	offsetTimeline *timeline.OffsetTimeline
 	lock           sync.RWMutex
 	log            *zap.SugaredLogger
+}
+
+func (p *ProcessorToFetch) Entity() ProcessorEntitier {
+	return p.entity
+}
+
+func (p *ProcessorToFetch) OffsetTimeline() *timeline.OffsetTimeline {
+	return p.offsetTimeline
 }
 
 func (p *ProcessorToFetch) String() string {
@@ -62,12 +70,12 @@ func (p *ProcessorToFetch) String() string {
 }
 
 // NewProcessorToFetch creates ProcessorToFetch.
-func NewProcessorToFetch(ctx context.Context, processor processor.ProcessorEntitier, capacity int) *ProcessorToFetch {
+func NewProcessorToFetch(ctx context.Context, processor ProcessorEntitier, capacity int) *ProcessorToFetch {
 	p := &ProcessorToFetch{
 		ctx:            ctx,
 		entity:         processor,
 		status:         _active,
-		offsetTimeline: NewOffsetTimeline(ctx, capacity),
+		offsetTimeline: timeline.NewOffsetTimeline(ctx, capacity),
 		log:            logging.FromContext(ctx),
 	}
 	return p
