+++ pkg/watermark/timeline/offset_timeline_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fetch
+package timeline
 
 import (
 	"context"
